Really big restructuring (see details)
* All files have been moved to the root module for convenience. It gave a
lot of trouble when restructuring the code and the project isn't really that
big anyway.

* The "Emulator" has been separated from the "GB". The rationale behind this
is that the debugging functionality shouldn't be part of the system itself.
When the time comes (hopefully soon) that it can show an image or play sounds,
the Emulator should be the one responsible for managing i/o.